parser/requirements.txt
alembic==1.6.5
astroid==2.4.2
attrs==20.1.0
bidict==0.21.2
boltons==21.0.0
certifi==2021.5.30
charset-normalizer==2.0.3
colander==0.9.9
Flask==0.10.1
Flask-Alembic==2.0.1
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==1.7.0
iniconfig==1.0.1
iso8601==0.1.8
isort==5.4.2
itsdangerous==1.1.0
Jinja2==2.11.2
kafka-python==2.0.2
lark-parser==0.9.0
lazy-object-proxy==1.4.3
Mako==1.1.4
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==4.2.0
python-socketio==5.4.0
redis==2.8.0
requests==2.26.0
six==1.15.0
SQLAlchemy==1.4.20
toml==0.10.1
translationstring==1.4
typed-ast==1.4.1
urllib3==1.26.6
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
